requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
openai==1.12.0
aioredis==2.0.1
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0
alembic==1.13.1
sqlalchemy==2.0.27
pytest==8.0.1
pytest-asyncio==0.23.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
qdrant-client==1.7.3
asyncpg==0.29.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0